nova-llm-agent/requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.1
pinecone-client==3.0.0
openai==1.6.1 